requirements.txt
bcrypt>=3.1.7
Faker>=2.0.3
Flask>=1.1.1
flask-restplus>=0.13.0
numpy>=1.17.3
pandas>=0.25.3
pyjwt>=1.7.1
scikit-learn>=0.21.3
SQLAlchemy>=1.3.10
